+++ packages/orq-rc/src/lib/config.ts
@@ -58,7 +58,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.11.0-rc.21",
+  sdkVersion: "4.11.0-rc.22",
   genVersion: "2.904.2",
-  userAgent: "speakeasy-sdk/typescript 4.11.0-rc.21 2.904.2 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.11.0-rc.22 2.904.2 2.0 @orq-ai/node",
 } as const;