+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3861,7 +3861,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KV5DTN0KTJFDADH1QPESQY6B"),
+  id: z.string().default("01KV5QJJ6FFSA66CSTH4WYSXVR"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),