+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-15T10:38:53.975Z",
+    "2026-06-15T13:29:06.223Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {