+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -152,7 +152,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KV5DTNB8FM7AMHW39HWPX50W"),
+  _id: z.string().default("01KV5QJJM1WAKN8ZZ5BWZ50Y9N"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: UpdateDatasourceStatus$inboundSchema,